Deleting a Post no longer does a cascade delete which removes the User, User Id and Comments, it not only removes the post and its comments

+++ src/main/java/com/project/service/PostServiceImpl.java
@@ -10,6 +10,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.List;
 
 @Service
@@ -51,9 +52,18 @@ public class PostServiceImpl implements PostService {
     }
 
     @Override
+    @Transactional
     public HttpStatus deleteById(Long postId){
-        postRepository.deleteById(postId);
-        return HttpStatus.OK;
+        Authentication auth = authenticationImpl.getAuthentication();
+        Long userId = userRepository.findByUsername(auth.getName()).getId();
+        Long postUserId = postRepository.findPostById(postId).getUser().getId();
+        if (userId.equals(postUserId)){
+            postRepository.deleteById(postId);
+            return HttpStatus.OK;
+        }
+        return HttpStatus.FORBIDDEN;
+
+
     }
 
 }
